multi: add dev build tag
