--- ui/materialplus/modal.go
@@ -1,40 +1,0 @@
-package materialplus
-
-import (
-	"image/color"
-
-	"gioui.org/layout"
-	"gioui.org/unit"
-)
-
-const (
-	heightPercentage = 80 // percentage of window height the modal content takes
-)
-
-// Modal renders a modal instance to screen
-func (t *Theme) Modal(gtx *layout.Context, renderFunc func()) {
-	overlayColor := color.RGBA{255, 255, 255, 255}
-	overlayColor.A = 200
-
-	Fill(gtx, overlayColor, gtx.Constraints.Width.Max, gtx.Constraints.Height.Max)
-
-	layout.Stack{}.Layout(gtx,
-		layout.Expanded(func() {
-			gtx.Constraints.Height.Min = (heightPercentage / 100) * gtx.Constraints.Height.Max
-		}),
-		layout.Stacked(func() {
-			inset := layout.Inset{
-				Top: unit.Dp(50),
-			}
-			inset.Layout(gtx, func() {
-				Fill(gtx, color.RGBA{}, gtx.Constraints.Width.Max, gtx.Constraints.Height.Max)
-				inset := layout.Inset{
-					Top:   unit.Dp(7),
-					Left:  unit.Dp(25),
-					Right: unit.Dp(25),
-				}
-				inset.Layout(gtx, renderFunc)
-			})
-		}),
-	)
-}
