+++ ui/window/loading.go
@@ -1,0 +1,7 @@
+package window
+
+import "gioui.org/layout"
+
+func (win *Window) Loading(gtx *layout.Context) {
+
+}
