+++ ui/window/window.go
@@ -2,11 +2,10 @@ package window
 
 import (
 	"fmt"
-	"time"
 
 	"gioui.org/app"
-	"gioui.org/io/system"
 	"gioui.org/layout"
+	"gioui.org/widget"
 
 	"github.com/raedahgroup/godcr-gio/ui"
 	"github.com/raedahgroup/godcr-gio/ui/materialplus"
@@ -25,6 +24,9 @@ type Window struct {
 	states     map[string]interface{}
 	uiEvents   chan interface{}
 	walletInfo *wallet.MultiWalletInfo
+	buttons    struct {
+		deleteWallet, cancelDialog, confirmDialog *widget.Button
+	}
 }
 
 // CreateWindow creates and initializes a new window with start
@@ -46,7 +48,7 @@ func CreateWindow(start string, wal *wallet.Wallet) (*Window, error) {
 	pages[page.LandingID] = new(page.Landing)
 	pages[page.LoadingID] = new(page.Loading)
 	pages[page.WalletsID] = new(page.Wallets)
-	pages[page.UITestID] = new(page.UITest)
+	//pages[page.UITestID] = new(page.UITest)
 
 	win.walletInfo = new(wallet.MultiWalletInfo)
 	win.states[page.StateWalletInfo] = win.walletInfo
@@ -62,66 +64,6 @@ func CreateWindow(start string, wal *wallet.Wallet) (*Window, error) {
 	win.pages = pages
 	win.wallet = wal
 	return win, nil
-}
-
-// Loop runs main event handling and page rendering loop
-func (win *Window) Loop(shutdown chan int) {
-	for {
-		select {
-		case e := <-win.uiEvents:
-			switch evt := e.(type) {
-			case page.EventNav:
-				win.current = evt.Next
-			case error:
-				// TODO: display error
-			}
-			win.window.Invalidate()
-		case e := <-win.wallet.Send:
-			log.Debugf("Recieved event %+v", e)
-			if e.Err != nil {
-				win.states[page.StateError] = e.Err
-				win.window.Invalidate()
-				break
-			}
-			switch evt := e.Resp.(type) {
-			case *wallet.LoadedWallets:
-				win.wallet.GetMultiWalletInfo()
-				if evt.Count == 0 {
-					win.current = page.LandingID
-				} else {
-					win.current = page.WalletsID
-				}
-			case *wallet.MultiWalletInfo:
-				*win.walletInfo = *evt
-			default:
-				win.updateState(e.Resp)
-			}
-			// set error if it exists
-			if e.Err != nil {
-				win.states[page.StateError] = e.Err
-			}
-			win.window.Invalidate()
-		case e := <-win.window.Events():
-			switch evt := e.(type) {
-			case system.DestroyEvent:
-				close(shutdown)
-				return
-			case system.FrameEvent:
-				win.gtx.Reset(evt.Config, evt.Size)
-				start := time.Now()
-				pageEvt := win.pages[win.current].Draw(win.gtx)
-				log.Tracef("Page {%s} rendered in %v", win.current, time.Since(start))
-				if pageEvt != nil {
-					win.uiEvents <- pageEvt
-				}
-				evt.Frame(win.gtx.Ops)
-			case nil:
-				// Ignore
-			default:
-				log.Tracef("Unhandled window event %+v\n", e)
-			}
-		}
-	}
 }
 
 // updateState checks for the event type that is passed as an argument and updates its
